chore: refactor CTR content across service pages and components
