+++ src/components/global/Footer.astro
@@ -173,15 +173,15 @@ const currentPath = currentPathRaw !== "/" ? currentPathRaw : "/";
             </a>
           </p>
         </div>
-        <div class="md:text-right md:w-fit">
+        <div class="w-fit md:text-right md:ml-auto">
           <Button
             href="#"
             type="button"
             variant="primary"
-            className=" flex items-center gap-2"
+            className="flex items-center gap-2"
             id="back-to-top"
           >
-            <ArrowUpIcon />
+            <ArrowUpIcon class="w-5 h-5 shrink-0" />
             Volver al inicio
           </Button>
         </div>
