+++ src/components/global/NavbarAlt.astro
@@ -222,11 +222,12 @@ const currentPath = currentPathRaw !== "/" ? currentPathRaw : "/";
       </div>
     </div>
   </nav>
-  <!-- Menu Button and Theme Toggle -->
-  <div class="absolute top-6 right-6 lg:right-16 items-center flex gap-4 z-50">
-    <div class="lg:hidden z-50 relative opacity-100">
-      <ThemeToggle />
-    </div>
+  <!-- Theme Toggle (mobile only — desktop version lives inside #visible-menu) -->
+  <div class="lg:hidden absolute top-5 right-16 z-50">
+    <ThemeToggle />
+  </div>
+  <!-- Menu Button -->
+  <div class="absolute top-6 right-6 lg:right-16 z-50">
     <button
       type="button"
       id="btn-open-close"
