+++ src/components/section/ourServices/AstroVsWordpressTable.astro
@@ -17,7 +17,7 @@ const lastRow = rows.length - 1;
 ---
 
 <section class="w-full">
-  <div class="max-w-5xl mx-auto px-4 py-10 lg:py-24">
+  <div class="max-w-5xl mx-auto px-4 pt-10 pb-6 lg:pt-24 lg:pb-8">
 
     <!-- Header: Title and description only -->
     <div class="mb-8 lg:mb-10">
