+++ src/components/section/ourServices/ProcessSteps.astro
@@ -1,5 +1,6 @@
 ---
 import type { Service } from "@/config/services";
+import MouseDown from "@/assets/icons/mousedwn.svg";
 
 interface Props {
   service: Service;
@@ -8,10 +9,30 @@ interface Props {
 const { service } = Astro.props;
 
 const panelAccents = [
-  { label: "accentGold", border: "border-l-accentGold", text: "text-accentGold", line: "bg-accentGold" },
-  { label: "sandRose", border: "border-l-sandRose", text: "text-sandRose", line: "bg-sandRose" },
-  { label: "salviaGreen", border: "border-l-salviaGreen", text: "text-salviaGreen", line: "bg-salviaGreen" },
-  { label: "accentGold", border: "border-l-accentGold", text: "text-accentGold", line: "bg-accentGold" },
+  {
+    label: "accentGold",
+    border: "border-l-accentGold",
+    text: "text-accentGold",
+    line: "bg-accentGold",
+  },
+  {
+    label: "sandRose",
+    border: "border-l-sandRose",
+    text: "text-sandRose",
+    line: "bg-sandRose",
+  },
+  {
+    label: "salviaGreen",
+    border: "border-l-salviaGreen",
+    text: "text-salviaGreen",
+    line: "bg-salviaGreen",
+  },
+  {
+    label: "accentGold",
+    border: "border-l-accentGold",
+    text: "text-accentGold",
+    line: "bg-accentGold",
+  },
 ] as const;
 
 const animationClasses = ["green", "orange", "purple", "blue"] as const;
@@ -19,9 +40,11 @@ const animationClasses = ["green", "orange", "purple", "blue"] as const;
 
 <section id="process" class="relative w-full">
   <div
-    class="description flex flex-col items-center justify-center w-full bg-smokyBlack px-6 py-16 lg:py-20 lg:min-h-screen lg:px-10"
+    class="description flex flex-col items-center justify-center w-full bg-smokyBlack px-6 py-16 lg:py-28 lg:px-10"
   >
-    <p class="panel-label text-sm uppercase tracking-widest text-accentGold/70 mb-4">
+    <p
+      class="panel-label text-sm uppercase tracking-widest text-accentGold/70 mb-4"
+    >
       Nuestro proceso
     </p>
     <h2
@@ -30,9 +53,13 @@ const animationClasses = ["green", "orange", "purple", "blue"] as const;
       <span class="block">De la primera llamada</span>
       <span class="block">a tu sitio web en vivo</span>
     </h2>
-    <p class="panel-sub mt-6 text-lg text-softBeige/70 max-w-xl text-center leading-relaxed">
-      Sin sorpresas. Sin cambios de alcance a mitad del proyecto. Te decimos exactamente qué pasa en cada etapa y qué recibes.
+    <p
+      class="panel-sub mt-6 text-lg text-softBeige/70 max-w-xl text-center leading-relaxed"
+    >
+      Sin sorpresas. Sin cambios de alcance a mitad del proyecto. Te decimos
+      exactamente qué pasa en cada etapa y qué recibes.
     </p>
+    <MouseDown class="mt-8 w-10 h-10 animate-bounce text-accentGold" />
   </div>
 
   <!-- Mobile/tablet: normal scroll flow, each step stacked -->
@@ -46,7 +73,9 @@ const animationClasses = ["green", "orange", "purple", "blue"] as const;
           <article class="mobile-step bg-smokyBlack px-6 py-14 md:py-20 flex items-center justify-center">
             <div class="w-full max-w-lg mx-auto">
               <div class={`border-l-2 ${accent.border} pl-6`}>
-                <p class={`text-xs uppercase tracking-widest ${accent.text} mb-4 font-semibold`}>
+                <p
+                  class={`text-xs uppercase tracking-widest ${accent.text} mb-4 font-semibold`}
+                >
                   Paso {stepNum}
                 </p>
 
@@ -58,7 +87,7 @@ const animationClasses = ["green", "orange", "purple", "blue"] as const;
                   {step.description}
                 </p>
 
-                <div class={`mt-8 h-px w-24 ${accent.line} opacity-60`}></div>
+                <div class={`mt-8 h-px w-24 ${accent.line} opacity-60`} />
               </div>
             </div>
           </article>
@@ -84,13 +113,19 @@ const animationClasses = ["green", "orange", "purple", "blue"] as const;
           >
             <div class="relative z-10 w-full max-w-5xl mx-auto px-6 lg:px-16">
               <div class="hidden lg:block absolute -top-40 -left-8 select-none pointer-events-none">
-                <span class={`panel-label text-watermark font-bold leading-none ${accent.text} opacity-5`}>
+                <span
+                  class={`panel-label text-watermark font-bold leading-none ${accent.text} opacity-5`}
+                >
                   {stepNum}
                 </span>
               </div>
 
-              <div class={`border-l-2 lg:border-l-4 ${accent.border} pl-6 lg:pl-10`}>
-                <p class={`panel-label text-xs uppercase tracking-widest ${accent.text} mb-4 font-semibold`}>
+              <div
+                class={`border-l-2 lg:border-l-4 ${accent.border} pl-6 lg:pl-10`}
+              >
+                <p
+                  class={`panel-label text-xs uppercase tracking-widest ${accent.text} mb-4 font-semibold`}
+                >
                   Paso {stepNum}
                 </p>
 
@@ -102,7 +137,9 @@ const animationClasses = ["green", "orange", "purple", "blue"] as const;
                   {step.description}
                 </p>
 
-                <div class={`mt-8 h-px w-24 lg:w-32 ${accent.line} opacity-60`}></div>
+                <div
+                  class={`mt-8 h-px w-24 lg:w-32 ${accent.line} opacity-60`}
+                />
               </div>
             </div>
           </section>
@@ -113,7 +150,7 @@ const animationClasses = ["green", "orange", "purple", "blue"] as const;
 </section>
 
 <script>
-  import { pinSlideInAnimation } from "@/utils/scripts/perPageAnimations/servicesAnimations/servicesStepsProcessAnimation2.js";
+  import { pinSlideInAnimation } from "@/utils/scripts/perPageAnimations/servicesAnimations/ServicesPinSlideAnimation.js";
 
   pinSlideInAnimation();
 </script>
