+++ src/components/section/ourServices/ServiceProblemSection.astro
@@ -2,55 +2,120 @@
 import { Image } from "astro:assets";
 
 interface Props {
-  title: string,
-  image: ImageMetadata,
-  imageAlt: string,
-  introText?: string,
-  problems: string[],
-  imagePosition?: "left" | "right"
+  title: string;
+  image: ImageMetadata;
+  imageAlt: string;
+  introText?: string;
+  problems: string[];
+  imagePosition?: "left" | "right";
 }
 
-const { title, image, imageAlt, introText, problems, imagePosition } = Astro.props;
+const {
+  title,
+  image,
+  imageAlt,
+  introText,
+  problems,
+  imagePosition = "left",
+} = Astro.props;
 ---
 
-<section id="problem-section" class="max-w-6xl mx-auto py-10 lg:py-20">
-    <div class="problem-heading mb-8 opacity-0">
-      <h2 class="text-center text-smokyBlack dark:text-softWhite text-[min(10vw,42px)]">
-      {title}
-    </h2>
-    </div>
-    <div class="grid grid-cols-1 md:grid-cols-2 gap-4 md:gap-8 items-center">
-      <div id="problem-image" class="w-full flex items-center justify-center opacity-0">
-        <Image
-          src={image}
-          alt={imageAlt}
-          priority={true}
-          class="object-cover w-full max-w-sm "
+<section id="problem-section" class="relative overflow-hidden py-16 lg:py-28">
+  <!-- Ambient glows -->
+  <div
+    class="pointer-events-none absolute -top-32 -right-32 h-96 w-96 rounded-full bg-accentGold/5 blur-3xl"
+    aria-hidden="true"
+  >
+  </div>
+  <div
+    class="pointer-events-none absolute -bottom-8 -left-16 h-72 w-72 rounded-full bg-sandRose/5 blur-3xl"
+    aria-hidden="true"
+  >
+  </div>
 
-        />
+  <div class="mx-auto max-w-6xl px-4 md:px-8">
+    <!-- Section heading -->
+    <div class="problem-heading mb-12 opacity-0 text-center md:mb-20">
+      <p
+        class="mb-3 text-sm font-semibold uppercase tracking-widest text-accentGold"
+      >
+        El Problema
+      </p>
+      <h2
+        class="mx-auto max-w-3xl text-4xl leading-tight text-smokyBlack dark:text-softWhite md:text-5xl"
+      >
+        {title}
+      </h2>
+      <div class="mx-auto mt-6 h-px w-16 bg-accentGold"></div>
+    </div>
+
+    <!-- Content grid -->
+    <div class="grid grid-cols-1 items-center gap-10 md:grid-cols-2 md:gap-16">
+      <!-- Image panel -->
+      <div
+        id="problem-image"
+        class={`relative opacity-0 ${imagePosition === "right" ? "md:order-2" : "md:order-1"}`}
+      >
+        <!-- Offset decorative border -->
+        <div
+          class="absolute -inset-3 rounded-3xl border border-accentGold/20"
+          aria-hidden="true"
+        >
+        </div>
+        <!-- Corner ambient glow -->
+        <div
+          class="absolute -bottom-4 -right-4 h-28 w-28 rounded-full bg-accentGold/10 blur-2xl"
+          aria-hidden="true"
+        >
+        </div>
+        <div
+          class="relative overflow-hidden rounded-2xl shadow-2xl shadow-smokyBlack/10 dark:shadow-softWhite/5"
+        >
+          <Image
+            src={image}
+            alt={imageAlt}
+            priority
+            class="w-full h-auto object-cover"
+          />
+        </div>
       </div>
-      <div id="problem-hurts" class="w-full flex flex-col justify-center">
-        {introText && (
-          <h3 class="text-smokyBlack dark:text-softWhite mb-4 text-center md:text-left text-xl md:text-2xl font-semibold leading-tight opacity-0">
-            {introText}
-          </h3>
-          <ul class="flex flex-col gap-2">
-            {problems.map((hurt) => (
-              <li class="hurts text-smokyBlack dark:text-softWhite text-lg flex items-center gap-2 px-2 opacity-0">
-                <span class="w-3 h-3 bg-accentGold rounded-full shrink-0" />
-                {hurt}
+
+      <!-- Pain points panel -->
+      <div
+        id="problem-hurts"
+        class={`flex flex-col gap-5 ${imagePosition === "right" ? "md:order-1" : "md:order-2"}`}
+      >
+        {
+          introText && (
+            <h4 class="border-l-2 border-accentGold pl-4 text-xl font-semibold leading-snug text-smokyBlack opacity-0 dark:text-softWhite md:text-2xl">
+              {introText}
+            </h4>
+          )
+        }
+
+        <ul class="mt-1 flex flex-col gap-3">
+          {
+            problems.map((hurt, index) => (
+              <li class="hurts opacity-0">
+                <div class="group flex items-start gap-4 rounded-xl border border-smokyBlack/10 bg-softBeige/50 p-4 transition-colors duration-200 hover:border-accentGold/40 dark:border-softWhite/10 dark:bg-softWhite/5">
+                  <span class="flex h-8 w-8 shrink-0 items-center justify-center rounded-full border border-accentGold/40 bg-accentGold/15 text-sm font-bold text-smokyBlack dark:text-softWhite">
+                    {index + 1}
+                  </span>
+                  <span class="pt-1 text-base leading-relaxed text-smokyBlack dark:text-softWhite">
+                    {hurt}
+                  </span>
+                </div>
               </li>
-            ))}
-          </ul>
-        )}
+            ))
+          }
+        </ul>
       </div>
     </div>
-  </section>
+  </div>
+</section>
 
-  <script>
-    import { gsapProblemAnimation } from "@/utils/scripts/perPageAnimations/servicesAnimations/servicesProblemAnimation";
-    
-    document.addEventListener("DOMContentLoaded", () => {
-      gsapProblemAnimation();
-    });
-  </script>
+<script>
+  import { gsapProblemAnimation } from "@/utils/scripts/perPageAnimations/servicesAnimations/servicesProblemAnimation";
+
+  gsapProblemAnimation();
+</script>
