+++ src/components/section/perPage/AboutFoundationsAlt.astro
@@ -35,31 +35,41 @@ const foundations = [
 ];
 ---
 
-<section class="foundations-alt-section w-full py-10 lg:py-16 px-4 overflow-visible">
+<section
+  class="foundations-alt-section w-full py-10 lg:py-16 px-4 overflow-visible"
+>
   <!-- Heading -->
   <div class="max-w-3xl mx-auto text-center mb-8 lg:mb-12">
-    <h2 class="text-[min(10vw,56px)] text-smokyBlack dark:text-softWhite tracking-tight leading-tight invisible">
+    <h2
+      class="text-[min(10vw,56px)] text-smokyBlack dark:text-softWhite tracking-tight leading-tight invisible"
+    >
       Nuestros <em class="not-italic text-accentGold">Principios</em>
     </h2>
-    <p class="mt-3 text-sm lg:text-base text-smokyBlack/60 dark:text-softWhite/55 max-w-xl mx-auto leading-relaxed">
+    <p
+      class="mt-3 text-sm lg:text-base text-smokyBlack/60 dark:text-softWhite/55 max-w-xl mx-auto leading-relaxed"
+    >
       Lo que nos guía en cada proyecto, con cada cliente, en cada decisión.
     </p>
   </div>
 
   <!-- Card grid -->
-  <div class="max-w-4xl mx-auto grid grid-cols-1 md:grid-cols-3 gap-6 md:gap-4 lg:gap-6 items-start">
-    {foundations.map((f, i) => (
-      <FoundationCard
-        title={f.title}
-        text={f.text}
-        Icon={f.Icon}
-        cardBg={f.cardBg}
-        blobColor={f.blobColor}
-        iconColor={f.iconColor}
-        rotate={f.rotate}
-        offset={i === 1 ? "md:mt-8" : ""}
-      />
-    ))}
+  <div
+    class="max-w-4xl mx-auto grid grid-cols-1 md:grid-cols-3 gap-6 md:gap-4 lg:gap-6 items-start"
+  >
+    {
+      foundations.map((f, i) => (
+        <FoundationCard
+          title={f.title}
+          text={f.text}
+          Icon={f.Icon}
+          cardBg={f.cardBg}
+          blobColor={f.blobColor}
+          iconColor={f.iconColor}
+          rotate={f.rotate}
+          offset={i === 1 ? "md:mt-8" : ""}
+        />
+      ))
+    }
   </div>
 </section>
