+++ src/components/section/perPage/AboutServiceLinks.astro
@@ -49,7 +49,7 @@ const services = [
               {/* Ghost number */}
               <span
                 aria-hidden="true"
-                class="shrink-0 font-playfair text-[min(14vw,96px)] font-bold leading-none text-smokyBlack/8 dark:text-softWhite/8 select-none w-14 lg:w-28 text-right tabular-nums"
+                class="shrink-0 font-playfair text-[clamp(3rem,10vw,6rem)] font-bold leading-none text-smokyBlack/8 dark:text-softWhite/8 select-none w-16 md:w-24 lg:w-28 text-right tabular-nums"
               >
                 {service.number}
               </span>
