+++ src/components/ui/ServiceCard.astro
@@ -42,9 +42,7 @@ const { slug, image, alt, title, description, tags } = Astro.props;
           {title}
         </h3>
       </div>
-      <p
-        class="description text-sm text-deepplum dark:text-softWhite mt-2"
-      >
+      <p class="description text-sm text-deepplum dark:text-softWhite mt-2">
         {description}
       </p>
       <Button
@@ -52,7 +50,7 @@ const { slug, image, alt, title, description, tags } = Astro.props;
         className="mt-4 text-sm"
         variant="optional1"
       >
-        Ver más
+        Ver Servicio
       </Button>
     </div>
   </div>
