+++ src/content/works/buitragoyvillota.md
@@ -1,0 +1,15 @@
+---
+pubDate: 2026-06-05
+updatedDate: 2026-06-05
+title: Buitrago y Villota
+description: "Desarrollo web y estrategia SEO para Buitrago y Villota"
+image: "./img/buitragoyvillota.webp"
+link: "https://buitragoyvillota.com/"
+stack: ["Astro", "TypeScript", "TailwindCSS", "SEO"]
+tags: ["Branding", "Desarrollo Web", "Optimización SEO"]
+---
+
+# Key project features:
+
+- Desarrollo de sitio web corporativo profesional
+- Estrategia e implementación de posicionamiento SEO
