+++ src/content/works/elianaalvarezpsicologa.md
@@ -1,0 +1,15 @@
+---
+pubDate: 2026-06-08
+updatedDate: 2026-06-08
+title: Eliana Alvarez Psicologa
+description: "Rediseño web y estrategia SEO para Eliana Alvarez Psicologa"
+image: "./img/elianaalvarezpsicologa.webp"
+link: "https://elianaalvarezpsicologa.com/"
+stack: ["Astro", "TypeScript", "TailwindCSS", "SEO"]
+tags: ["Branding", "Rediseño Web", "Optimización SEO", "Migración de sitio wordpress a Astro"]
+---
+
+# Key project features:
+- Migración del sitio web de WordPress a Astro
+- Rediseño de sitio web corporativo profesional
+- Estrategia e implementación de posicionamiento SEO
