+++ src/content/works/recurrirabogados.md
@@ -1,0 +1,15 @@
+---
+pubDate: 2026-05-15
+updatedDate: 2026-05-15
+title: Recurrira Bogados
+description: "Desarrollo web y estrategia SEO para Recurrira Bogados"
+image: "./img/recurrirabogados.webp"
+link: "https://recurrirabogados.com/"
+stack: ["Astro", "TypeScript", "TailwindCSS", "SEO"]
+tags: ["Branding", "Desarrollo Web", "Optimización SEO"]
+---
+
+# Key project features:
+
+- Desarrollo de sitio web corporativo profesional
+- Estrategia e implementación de posicionamiento SEO
