+++ src/pages/nosotros.astro
@@ -14,7 +14,6 @@ import {
   COMPANY_INFO,
 } from "@/config/seo";
 
-
 const breadcrumbSchema = generateBreadcrumbSchema([
   { name: "Inicio", url: `${COMPANY_INFO.url}/` },
   { name: "Nosotros", url: `${COMPANY_INFO.url}/nosotros/` },
@@ -25,7 +24,7 @@ const authorSchemas = generateAllAuthorsSchemas();
 const nosotrosSEO = generatePageSEO({
   title: "Rocío Parra y Diego Rodríguez | Equipo Shine Agencia Colombia",
   description:
-    "Conoce a Rocío y Diego, el equipo detrás de Shine Agencia: más de 10 años en marketing digital, desarrollo web con Astro y ecommerce con Tienda Nube en Colombia.",
+    "Rocío y Diego fundadores Shine Agencia: más de 10 años en marketing digital, desarrollo web con Astro y tiendas online con Tienda Nube en Bogotá, Colombia.",
   path: "/nosotros",
 });
 ---
